[Refactor] 파일 분리 및 전역 변수 삭제
