+++ LuckyCardGame/LuckyCardGame/ConstantSize.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+enum ConstantSize {
+    static let totalHeight = UIScreen.main.bounds.height
+    static let totalWidth = UIScreen.main.bounds.width
+    static let padding = totalWidth / 30
+}
